Add files via upload
This script can analize all subdirectories in the root directory and publis enabled files and folders.
The folders will be published as a category and items of the menu
The video filse will be published as a video-card with video-player, picture and short definition
The audio files will be published as a audio-card with audio-player with short definition
The text files will be published as an article-card with picture and text

Usage: unzip this into root directory and put into the subdirectories your content. You can place the mix of the different types of the content files.  
For full properties of the video publish you must place in the one place 3 files with the same name: mp4, jgp and txt
For full  properties of the audio publish you must place in the one place 2 files with the same name: mp3 and txt 
For publish the gallery you must place the single images in any types jpg, png or gif including gif animations